crawler/requirements.txt
aiohttp>=3.9.0
asyncpg>=0.29.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pytz>=2023.3
python-slugify>=8.0.0
tenacity>=8.2.0
structlog>=23.2.0
click>=8.1.0
python-dotenv>=1.0.0
